General Discussion

Archimedean copulas are a large class of copulas that are defined as : ... [ details needed...]

In our implementation, we separate the generator, in the Generator type, from the copula itself. The easyest way to understand this orthogonal design is to look at the Clayton implementation.

We have the ClaytonGenerator that is defined as follows:

struct ClaytonGenerator{T} <: UnivariateGenerator
    θ::T
    function ClaytonGenerator(θ)
        if θ < -1
            throw(ArgumentError("Theta must be greater than -1"))
        elseif θ == -1
            return WGenerator()
        elseif θ == 0
            return IndependentGenerator()
        elseif θ == Inf
            return MGenerator()
        else
            return new{typeof(θ)}(θ)
        end
    end
end
max_monotony(G::ClaytonGenerator) = G.θ >= 0 ? Inf : Int(floor(1 - 1/G.θ))
ϕ(  G::ClaytonGenerator, t) = max(1+G.θ*t,zero(t))^(-1/G.θ)
ϕ⁻¹(G::ClaytonGenerator, t) = (t^(-G.θ)-1)/G.θ
# ϕ⁽¹⁾(G::ClaytonGenerator, t) =  First derivative of ϕ
# ϕ⁽ᵏ⁾(G::ClaytonGenerator, k, t) = kth derivative of ϕ
τ(G::ClaytonGenerator) = ifelse(isfinite(G.θ), G.θ/(G.θ+2), 1)
τ⁻¹(::Type{T},τ) where T<:ClaytonGenerator = ifelse(τ == 1,Inf,2τ/(1-τ))
williamson_dist(G::ClaytonGenerator, d) = G.θ >= 0 ? WilliamsonFromFrailty(Distributions.Gamma(1/G.θ,G.θ),d) : ClaytonWilliamsonDistribution(G.θ,d)

And then the ClaytonCopula{d,T} is simply an alias for ArchimedeanCopula{d,ClaytonGenerator{T}}:

const ClaytonCopula{d,T} = ArchimedeanCopula{d,ClaytonGenerator{T}}
ClaytonCopula(d,θ) = ArchimedeanCopula(d,ClaytonGenerator(θ))
generatorof(::Type{ClaytonCopula}) = ClaytonGenerator

Not everything is necessary however, depending on the features you need. Indeed, the Archimedean API is modular:

Only ϕ and max_monotony are really required. Indeed, all other functions have generic fallbacks. In particular, the williamson_dist has a generic fallback that uses WilliamsonTransforms.jl for any generator. Note however that providing the williamson_dist yourself if you know it will allow sampling to be an order of magnitude faster: see how in the ClaytonCopula case we provided a sampler that is different for positive parameters (the generator is completely monotonous, and thus the frailty distribution is known, so we used WilliamsonFromFrailty), and the negative dependent cases, where only the CDF is known and implemented in ClaytonWilliamsonDistribution.

To evaluate the cdf and (log-)density in any dimension, only ϕ and ϕ⁻¹ are needed.

Currently, to fit the copula τ⁻¹ is needed as we use the inverse tau moment method. But we plan on also implementing inverse rho and MLE (density needed).

Note that the generator ϕ follows the convention ϕ(0)=1, while others (e.g., https://en.wikipedia.org/wiki/Copula(probabilitytheory)#Archimedean_copulas) use ϕ⁻¹ as the generator.

Copulas.ArchimedeanCopula — Type

ArchimedeanCopula{d, TG}

Fields: - G::TG : the generator <: Generator.

Constructor:

ArchimedeanCopula(d::Int,G::Generator)

For some archimedean generator G::Generator and some dimenson d, this class models the archimedean copula wich has this generator. More formally, a $d$-monotone archimedean generator is a function $\phi$ on $\mathbb R_+$ that has these three properties:

$\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

In the code, this function is implemented as ϕ(G::Generator,t), and its max monotony d is given by max_monotony(G). The corresponding d-variate archimedean copula writes:

\[C(\mathbf u) = \phi^{-1}\left(\sum_{i=1}^d \phi(u_i)\right)\]

It can be sampled through A Radial simplex decomposition as follows: If $\mathbf U \sim C$, then $U \equal R \mathbf S$ for a random vector $\mathbf S \sim$ Dirichlet(ones(d)), that is uniformity on the d-variate simplex, and a non-negative random variable $R$ that is the Williamson d-transform of \phi. The density, cdf, kendall tau, and may other properties of the Archimedean copula are then directly deduced from this relationship.

There exists several known parametric generators that are implement in the package. For every NamedGenerator <: Generator implemented in the package, we provide a type alias `NamedCopula{d,...} = ArchimedeanCopula{d,NamedGenerator{...}} to be able to manipulate the archimedean copulas without too much hassle for known and usefull special cases.

A generic archimdeanc copula can be constructed as follows:

struct MyGenerator <: Generator end
ϕ(G::MyGenerator,t) = exp(-t) # your archimedean generator, can be any d-monotonous function.
max_monotony(G::MyGenerator) = Inf # could depend on generators parameters. 
C = ArchimedeanCopula(d,MyGenerator())

The obtained model can be used as follows:

spl = rand(C,1000)   # sampling
cdf(C,spl)           # cdf
pdf(C,spl)           # pdf
loglikelihood(C,spl) # llh

The following functions have defaults but can be overridden for performance:

ϕ⁻¹(C::MyArchimedean, t)        # Inverse of ϕ
ϕ⁽¹⁾(C::MyArchimedean, t)       # first defrivative of ϕ
ϕ⁽ᵈ⁾(C::MyArchimedean,t)        # dth defrivative of ϕ
τ(C::MyArchimedean)             # Kendall tau
τ⁻¹(::Type{MyArchimedean},τ) =  # Inverse kendall tau
fit(::Type{MyArchimedean},data) # fitting.

As a bonus, If you know the williamson d-transform of your generator and not your generator itself, you may take a look at WilliamsonGenerator that implements them. If you rather know the frailty distribution, take a look at WilliamsonFromFrailty.

References: Williamson, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581 McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.

Implement your own

If you think that the WilliamsonCopula interface is too barebone and does not provide you with enough flexibility in your modeling of an archimedean copula, you might be intersted in the possiiblity to directly subtype ArchimedeanCopula and implement your own. This is actually a fairly easy process and you only need to implement a few functions. Let's here together try to reimplement come archimedean copula with the follçowing generator:

\[my_generator\]

(describe the process...)

struct MyAC{d,T} <: ArchimedeanCopula{d}
    par::T
end
ϕ(C::MyAC{d},x)
ϕ⁻¹(C::MyAC{d},x)
τ(C::MyAC{d})
τ⁻¹(::MyAC{d},τ)
williamson_dist(C::MyAC{d})