Archimedean Copulas

Details about the Archimedean Copulas

Specific instances

Copulas.IndependentCopula — Type

IndependentCopula{d,T}

Constructor

IndependentCopula(d, θ)

The Independant Copula in dimension $d$ is the simplest copula, that has the form :

\[C(\mathbf{x}) = \prod_{i=1}^d x_i.\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = \exp{-t}\]

Missing docstring.

Missing docstring for ClaytonCopulas. Check Documenter's build log for details.

Copulas.FranckCopula — Type

FranckCopula{d,T}

Fields:

θ::Real - parameter

Constructor

FranckCopula(d, θ)

The Franck copula in dimension $d$ is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

Copulas.AMHCopula — Type

AMHCopula{d,T}

Fields:

θ::Real - parameter

Constructor

AMHCopula(d, θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [0,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

Copulas.GumbelCopula — Type

GumbelCopula{d,T}

Fields:

θ::Real - parameter

Constructor

GumbelCopula(d, θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

Copulas.JoeCopula — Type

JoeCopula{d,T}

Fields:

θ::Real - parameter

Constructor

JoeCopula(d, θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]