Available Models

EmpiricalCopula

Copulas.EmpiricalCopula — Type

EmpiricalCopula{d,MT}

Fields:

u::MT - the matrix of observations.

Constructor

EmpiricalCopula(u;pseudos=true)

The EmpiricalCopula in dimension $d$ is parameterized by a pseudo-data matrix wich should have shape (d,N). Its expression is given as :

\[C(\mathbf x) = \frac{1}{N}\sum_{i=1}^n \mathbf 1_{\mathbf u_i \le \mathbf x}\]

This function is very practical, be be aware that this is not a true copula (since $\mathbf u$ are only pseudo-observations). The constructor allows you to pass dirctly pseudo-observations (the default) or will compute them for you. You can then compute the cdf of the copula, and sample it through the standard interface.

BernsteinCopula

Not implemeted yet!

Do not hesitate to come talk on our github !

CheckerboardCopula

Not implemeted yet!

Do not hesitate to come talk on our github !

BetaCopula

Not implemeted yet!

Do not hesitate to come talk on our github !