Extreme Values Copulas

AsymGalambosCopula

Copulas.AsymGalambosCopula — Type

AsymGalambosCopula{P}

Fields:

α::Real - Dependency parameter

θ::Vector - Asymmetry parameters (size 2)

Constructor

AsymGalambosCopula(α, θ)

The Asymmetric bivariate Galambos copula is parameterized by one dependence parameter $\alpha \in [0, \infty]$ and two asymmetry parameters $\theta_{i} \in [0,1], i=1,2$. It is an Extreme value copula with Pickands function:

\[\A(t) = 1 - ((\theta_1t)^{-\alpha}+(\theta_2(1-t))^{-\alpha})^{-\frac{1}{\alpha}} \]

It has a few special cases:

When α = 0, it is the Independent Copula

When θ₁ = θ₂ = 0, it is the Independent Copula

When θ₁ = θ₂ = 1, it is the Galambos Copula

References:

[56] Families of min-stable multivariate exponential and multivariate extreme value distributions. Statist. Probab, 1990.

AsymLogCopula

Copulas.AsymLogCopula — Type

AsymLogCopula{P}

Fields:

α::Real - Dependency parameter

θ::Vector - Asymmetry parameters (size 2)

Constructor

AsymLogCopula(α, θ)

The Asymmetric bivariate Logistic copula is parameterized by one dependence parameter $\alpha \in [1, \infty]$ and two asymmetry parameters $\theta_{i} \in [0,1], i=1,2$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = (\theta_1^{\alpha}(1-t)^{\alpha} + \theta_2^{\alpha}t^{\alpha})^{\frac{1}{\alpha}} + (\theta_1 - \theta_2)t + 1 - \theta_1\]

References:

[57] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

AsymMixedCopula

Copulas.AsymMixedCopula — Type

AsymMixedCopula{P}

Fields:

θ::Vector - parameters (size 2)

Constructor

AsymMixedCopula(θ)

The Asymmetric bivariate Mixed copula is parameterized by two parameters $\theta_{i}, i=1,2$ that must meet the following conditions:

θ₁ ≥ 0

θ₁+θ₂ ≤ 1

θ₁+2θ₂ ≤ 1

θ₁+3θ₂ ≥ 0

It is an Extreme value copula with Pickands dependence function:

\[A(t) = \theta_{2}t^3 + \theta_{1}t^2-(\theta_1+\theta_2)t+1\]

It has a few special cases:

When θ₁ = θ₂ = 0, it is the Independent Copula

When θ₂ = 0, it is the Mixed Copula

References:

[57] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

BC2Copula

Copulas.BC2Copula — Type

BC2Copula{P}

Fields:

- a::Real - parameter
- a::Real - parameter

Constructor

BC2Copula(a, b)

The bivariate BC2 copula is parameterized by two parameters $a,b \in [0,1]$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \max\{a t, b (1-t) \} + \max\{(1-a)t, (1-b)(1-t)\}\]

References:

[58] Mai, J. F., & Scherer, M. (2011). Bivariate extreme-value copulas with discrete Pickands dependence measure. Extremes, 14, 311-324. Springer, 2011.

CuadrasAugeCopula

Copulas.CuadrasAugeCopula — Type

CuadrasAugeCopula{P}

Fields:

- α::Real - parameter

Constructor

CuadrasAugeCopula(α)

The bivariate Cuadras Auge copula is parameterized by $\alpha \in [0,1]$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \max\{t, 1-t \} + (1-\theta)\max\{t, 1-t\}\]

References:

[59] Mai, J. F., & Scherer, M. (2012). Simulating copulas: stochastic models, sampling algorithms, and applications (Vol. 4). World Scientific.

GalambosCopula

Copulas.GalambosCopula — Type

GalambosCopula{P}

Fields:

- θ::Real - parameter

Constructor

GalambosCopula(θ)

The bivariate Galambos copula is parameterized by $\alpha \in [0,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = 1 - (t^{-\theta}+(1-t)^{-\theta})^{-\frac{1}{\theta}}\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

[60] Galambos, J. (1975). Order statistics of samples from multivariate distributions. Journal of the American Statistical Association, 70(351a), 674-680.

HuslerReissCopula

Copulas.HuslerReissCopula — Type

HuslerReissCopula{P}

Fields:

- θ::Real - parameter

Constructor

HuslerReissCopula(θ)

The bivariate Husler-Reiss copula is parameterized by $\theta \in [0,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = t\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{t}{1-t})) +(1-t)\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{1-t}{t}))\]

Where $\Phi$is the cumulative distribution function (CDF) of the standard normal distribution.

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

[61] Hüsler, J., & Reiss, R. D. (1989). Maxima of normal random vectors: between independence and complete dependence. Statistics & Probability Letters, 7(4), 283-286.

LogCopula

Copulas.LogCopula — Type

LogCopula{P}

Fields:

- θ::Real - parameter

Constructor

LogCopula(θ)

The bivariate Logistic copula (or Gumbel Copula) is parameterized by $\theta \in [1,\infty)$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = (t^{\theta}+(1-t)^{\theta})^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[57] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

MixedCopula

Copulas.MixedCopula — Type

MixedCopula{P}

Fields:

- θ::Real - parameter

Constructor

MixedCopula(θ)

The bivariate Mixed copula is parameterized by $\alpha \in [0,1]$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \theta t^2 - \theta t + 1\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

[57] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

MOCopula

Copulas.MOCopula — Type

MOCopula{P}

Fields:

- λ₁::Real - parameter
- λ₂::Real - parameter
- λ₁₂::Real - parameter

Constructor

MOCopula(θ)

The bivariate Marshall-Olkin copula is parameterized by $\lambda_i \in [0,\infty), i = 1, 2, \{1,2\}$. It is an Extreme value copula with Pickands dependence function:

\[A(t) = \frac{\lambda_1 (1-t)}{\lambda_1 + \lambda_{1,2}} + \frac{\lambda_2 t}{\lambda_2 + \lambda_{1,2}} + \lambda_{1,2}\max\left \{\frac{1-t}{\lambda_1 + \lambda_{1,2}}, \frac{t}{\lambda_2 + \lambda_{1,2}} \right \} \]

References:

[59] Mai, J. F., & Scherer, M. (2012). Simulating copulas: stochastic models, sampling algorithms, and applications (Vol. 4). World Scientific.

tEVCopula

Copulas.tEVCopula — Type

tEVCopula{P}

Fields: - ν::Real - paremeter - θ::Real - Parameter

Constructor

tEVCopula(ν, θ)

The bivariate extreme t copula is parameterized by $\nu \in [0,\infty)$ and \theta \in (-1,1]. It is an Extreme value copula with Pickands dependence function:

\[A(x) = xt_{\nu+1}(Z_x) +(1-x)t_{\nu+1}(Z_{1-x})\]

Where $t_{\nu + 1}$is the cumulative distribution function (CDF) of the standard t distribution with \nu + 1 degrees of freedom and

\[Z_x = \frac{(1+\nu)^{1/2}{\sqrt{1-\theta^2}}\left [ \left (\frac{x}{1-x} \right )^{1/\nu} - \theta \right ]\]

It has a few special cases:

When θ = 0, it is the Independent Copula

When θ = ∞, it is the M Copula (Upper Frechet-Hoeffding bound)

References:

[62] Nikoloulopoulos, A. K., Joe, H., & Li, H. (2009). Extreme value properties of multivariate t copulas. Extremes, 12, 129-148.

[56]

H. Joe. Families of min-stable multivariate exponential and multivariate extreme value distributions. Statistics & probability letters 9, 75–81 (1990).

[57]

J. A. Tawn. Bivariate extreme value theory: models and estimation. Biometrika 75, 397–415 (1988).

[58]

J.-F. Mai and M. Scherer. Bivariate extreme-value copulas with discrete Pickands dependence measure. Extremes 14, 311–324 (2011).

[59]

J.-F. Mai and M. Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 4 (World Scientific, 2012).

[60]

J. Galambos. Order statistics of samples from multivariate distributions. Journal of the American Statistical Association 70, 674–680 (1975).

[61]

J. Hüsler and R.-D. Reiss. Maxima of normal random vectors: between independence and complete dependence. Statistics & Probability Letters 7, 283–286 (1989).

[62]

A. K. Nikoloulopoulos, H. Joe and H. Li. Extreme value properties of multivariate t copulas. Extremes 12, 129–148 (2009).