Available Archimedean models

Independence Copula

Copulas.IndependentGenerator — Type

IndependentGenerator

Constructor

IndependentGenerator()
IndependentCopula(d)

The Independent Copula in dimension $d$ is the simplest copula, that has the form :

\[C(\mathbf{x}) = \prod_{i=1}^d x_i.\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = \exp{-t}\]

Frechet-Hoeffding bounds

The Frechet-Hoeffding bounds are availiable through the following functions:

Copulas.MGenerator — Type

MGenerator

Constructor

MGenerator()
MCopula(d)

The Upper Frechet-Hoeffding bound is the copula with the greatest value among all copulas. It correspond to comonotone random vectors.

For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas.

Copulas.WGenerator — Type

WGenerator

Constructor

WGenerator()
WCopula(d)

The Lower Frechet-Hoeffding bound is the copula with the lowest value among all copulas. Note that $W$ is only a proper copula when $d=2$, in greater dimensions it is still the (pointwise) lower bound, but not a copula anymore. For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas.

Clayton Copula

Copulas.ClaytonGenerator — Type

ClaytonGenerator{T}

Fields:

θ::Real - parameter

Constructor

ClaytonGenerator(θ)
ClaytonCopula(d,θ)

The Clayton copula in dimension $d$ is parameterized by $\theta \in [-1/(d-1),\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

It has a few special cases:

When θ = -1/(d-1), it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Frank Copula

Copulas.FrankGenerator — Type

FrankGenerator{T}

Fields:

θ::Real - parameter

Constructor

FrankGenerator(θ)
FrankCopula(d,θ)

The Frank copula in dimension $d$ is parameterized by $\theta \in [-\infty,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

It has a few special cases:

When θ = -∞, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Gumbel Copula

Copulas.GumbelGenerator — Type

GumbelGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelGenerator(θ)
GumbelCopula(d,θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Ali-Mikhail-Haq Copula

Copulas.AMHGenerator — Type

AMHGenerator{T}

Fields:

θ::Real - parameter

Constructor

AMHGenerator(θ)
AMHCopula(d,θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [-1,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

Joe Copula

Copulas.JoeGenerator — Type

JoeGenerator{T}

Fields:

θ::Real - parameter

Constructor

JoeGenerator(θ)
JoeCopula(d,θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Gumbel-Barnett Copula

Copulas.GumbelBarnettGenerator — Type

GumbelBarnettGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelBarnettGenerator(θ)
GumbelBarnettCopula(d,θ)

The Gumbel-Barnett copula is an archimdean copula with generator:

\[\phi(t) = \exp{θ^{-1}(1-e^{t})}, 0 \leq \theta \leq 1.\]

More details about Gumbel-Barnett copula are found in:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.437

It has a few special cases:

When θ = 0, it is the IndependentCopula

Inverse-Gaussian Copula

Copulas.InvGaussianGenerator — Type

InvGaussianGenerator{T}

Fields:

θ::Real - parameter

Constructor

InvGaussianGenerator(θ)
InvGaussianCopula(d,θ)

The Inverse Gaussian copula in dimension $d$ is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{\frac{(1-\sqrt{1+2θ^{2}t}}{θ}}.\]

More details about Inverse Gaussian Archimedean copula are found in :

Mai, Jan-Frederik, and Matthias Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 6. # N/A, 2017. Page 74.

It has a few special cases:

When θ = 0, it is the IndependentCopula