Archimedean Generators

WilliamsonGenerator

Copulas.WilliamsonGenerator — Type

WilliamsonGenerator{TX}
i𝒲{TX}

Fields:

X::TX – a random variable that represents its Williamson d-transform

d::Int – the dimension of the transformation.

Constructor

WilliamsonGenerator(X::Distributions.UnivariateDistribution, d)
i𝒲(X::Distributions.UnivariateDistribution,d)

The WilliamsonGenerator (alias i𝒲) allows to construct a d-monotonous archimedean generator from a positive random variable X::Distributions.UnivariateDistribution. The transformation, which is called the inverse Williamson transformation, is implemented in WilliamsonTransforms.jl.

For a univariate non-negative random variable $X$, with cumulative distribution function $F$ and an integer $d\ge 2$, the Williamson-d-transform of $X$ is the real function supported on $[0,\infty[$ given by:

\[\phi(t) = 𝒲_{d}(X)(t) = \int_{t}^{\infty} \left(1 - \frac{t}{x}\right)^{d-1} dF(x) = \mathbb E\left( (1 - \frac{t}{X})^{d-1}_+\right) \mathbb 1_{t > 0} + \left(1 - F(0)\right)\mathbb 1_{t <0}\]

This function has several properties:

We have that $\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

These properties makes this function what is called a d-monotone archimedean generator, able to generate archimedean copulas in dimensions up to $d$. Our implementation provides this through the Generator interface: the function $\phi$ can be accessed by

G = WilliamsonGenerator(X, d)
ϕ(G,t)

Note that you'll always have:

max_monotony(WilliamsonGenerator(X,d)) === d

References:

[22] Williamson, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581

[14] McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.

TiltedGenerator

Copulas.TiltedGenerator — Type

TiltedGenerator(G, p, sJ) <: Generator

Archimedean generator tilted by conditioning on p components fixed at values with cumulative generator sum sJ = ∑ ϕ⁻¹(u_j). It defines

ϕ_tilt(t) = ϕ^{(p)}(sJ + t) / ϕ^{(p)}(sJ)

and higher derivatives accordingly:

ϕ_tilt^{(k)}(t) = ϕ^{(k+p)}(sJ + t) / ϕ^{(p)}(sJ)

which yields the conditional copula within the Archimedean family for the remaining d-p variables. You will get a TiltedGenerator if you condition() an archimedean copula.

FrailtyGenerator

Copulas.FrailtyGenerator — Type

FrailtyGenerator<:AbstractFrailtyGenerator<:Generator

methods: - frailty(::FrailtyGenerator) gives the frailty - ϕ and the rest of generators are automatically defined from the frailty.

Constructor

FrailtyGenerator(D)

A Frailty generator can be defined by a positive random variable that happens to have a mgf() function to compute its moment generating function. The generator is simply:

\[\phi(t) = mgf(frailty(G), -t)\]

https://www.uni-ulm.de/fileadmin/websiteuniulm/mawi.inst.zawa/forschung/2009-08-16_hofert.pdf

References:

[57] M. Hoffert (2009). Efficiently sampling Archimedean copulas

ClaytonGenerator

Copulas.ClaytonGenerator — Type

ClaytonGenerator{T}

Fields:

θ::Real - parameter

Constructor

ClaytonGenerator(θ)
ClaytonCopula(d,θ)

The Clayton copula in dimension $d$ is parameterized by $\theta \in [-1/(d-1),\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

It has a few special cases:

When θ = -1/(d-1), it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

FrankGenerator

Copulas.FrankGenerator — Type

FrankGenerator{T}

Fields:

θ::Real - parameter

Constructor

FrankGenerator(θ)
FrankCopula(d,θ)

The Frank copula in dimension $d$ is parameterized by $\theta \in [-\infty,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

It has a few special cases:

When θ = -∞, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

GumbelGenerator

Copulas.GumbelGenerator — Type

GumbelGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelGenerator(θ)
GumbelCopula(d,θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

AMHGenerator

Copulas.AMHGenerator — Type

AMHGenerator{T}

Fields:

θ::Real - parameter

Constructor

AMHGenerator(θ)
AMHCopula(d,θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [-1,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

JoeGenerator

Copulas.JoeGenerator — Type

JoeGenerator{T}

Fields:

θ::Real - parameter

Constructor

JoeGenerator(θ)
JoeCopula(d,θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

GumbelBarnettGenerator

Copulas.GumbelBarnettGenerator — Type

GumbelBarnettGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelBarnettGenerator(θ)
GumbelBarnettCopula(d,θ)

The Gumbel-Barnett copula is an archimdean copula with generator:

\[\phi(t) = \exp{θ^{-1}(1-e^{t})}, 0 \leq \theta \leq 1.\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.437

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

InvGaussianGenerator

Copulas.InvGaussianGenerator — Type

InvGaussianGenerator{T}

Fields:

θ::Real - parameter

Constructor

InvGaussianGenerator(θ)
InvGaussianCopula(d,θ)

The Inverse Gaussian copula in dimension $d$ is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{\frac{1-\sqrt{1+2θ^{2}t}}{θ}}.\]

More details about Inverse Gaussian Archimedean copula are found in :

Mai, Jan-Frederik, and Matthias Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 6. # N/A, 2017. Page 74.

It has a few special cases:

When θ = 0, it is the IndependentCopula

References:

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006.

BB1Generator

Copulas.BB1Generator — Type

BB1Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB1Generator(θ, δ)
BB1Copula(d, θ, δ)

The BB1 copula is parameterized by $\theta \in (0,\infty)$ and $\delta \in [1, \infty)$. It is an Archimedean copula with generator :

\[\phi(t) = (1 + t^{\frac{1}{δ}})^{\frac{-1}{θ}},\]

It has a few special cases:

When δ = 1, it is the ClaytonCopula with parameter \theta.

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.190-192

BB2Generator

Copulas.BB2Generator — Type

BB2Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB2Generator(θ, δ)
BB2Copula(d, θ, δ)

The BB2 copula is parameterized by `\theta, \delta \in (0,\infty). It is an Archimedean copula with generator :

\[\phi(t) = [1 + \delta^{-1}log(1 + t)]^{-\frac{1}{\theta}},\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.193-194

BB3Generator

Copulas.BB3Generator — Type

BB3Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB3Generator(θ, δ)
BB3Copula(d, θ, δ)

The BB3 copula is parameterized by $\theta \in [1,\infty)$ and ``\delta \in (0,\infty). It is an Archimedean copula with generator :

\[\phi(t) = \exp(-[\delta^{-1}\log(1 + t)]^{\frac{1}{\theta}}),\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.195-196

BB6Generator

Copulas.BB6Generator — Type

BB6Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB6Generator(θ, δ)
BB6Copula(d, θ, δ)

The BB6 copula is parameterized by $\theta, \delta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - [1 - \exp(-t^{\frac{1}{\delta}})]^{\frac{1}{\theta}}\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.200-201

BB7Generator

Copulas.BB7Generator — Type

BB7Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB7Generator(θ, δ)
BB7Copula(d, θ, δ)

The BB7 copula is parameterized by $\theta \in [1,\infty)$ and $\delta \in (0, \infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - [1 - (1 + t)^{-\frac{1}{\delta}}]^{\frac{1}{\theta}},\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.202-203

BB8Generator

Copulas.BB8Generator — Type

BB8Generator{T}

Fields:

ϑ::Real - parameter

δ::Real - parameter

Constructor

BB8Generator(ϑ, δ)
BB8Copula(d, ϑ, δ)

The BB8 copula is parameterized by $\vartheta, \in [1,\infty)$ and $\delta \in (0, 1]$. It is an Archimedean copula with generator :

\[\phi(t) = \delta^{-1}[1 - (1 - \eta \exp(-t))^{\frac{1}{\vartheta}}],\]

where $\eta = 1 - (1 - \delta)^{\vartheta}.$

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.204-205

BB9Generator

Copulas.BB9Generator — Type

BB9Generator{T}

Fields:

ϑ::Real - parameter

δ::Real - parameter

Constructor

BB9Generator(ϑ, δ)
BB9Copula(d, ϑ, δ)

The BB9 copula is parameterized by $\vartheta, \in [1,\infty)$ and $\delta \in (0, \infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp(-(\delta^{-\vartheta} + t)^{\frac{1}{\vartheta}} + \delta^{-1}),\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.205-206

BB10Generator

Copulas.BB10Generator — Type

BB10Generator{T}

Fields:

θ::Real - parameter

δ::Real - parameter

Constructor

BB10Generator(θ, δ)
BB10Copula(d, θ, δ)

The BB10 copula is parameterized by $\theta, \in (0,\infty)$ and $\delta \in [0, 1]$. It is an Archimedean copula with generator :

\[\phi(t) = \Big(\tfrac{1-\delta}{e^{s}-\delta}\Big)^{1/\theta},\]

References:

[4] Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.206-207

[3]

R. B. Nelsen. An Introduction to Copulas. 2nd ed Edition, Springer Series in Statistics (Springer, New York, 2006).

[4]

H. Joe. Dependence Modeling with Copulas (CRC press, 2014).

[14]

A. J. McNeil and J. Nešlehová. Multivariate Archimedean Copulas, d -Monotone Functions and L1 -Norm Symmetric Distributions. The Annals of Statistics 37, 3059–3097 (2009).

[22]

R. E. Williamson. On multiply monotone functions and their laplace transforms (Mathematics Division, Office of Scientific Research, US Air Force, 1955).

[57]

M. Hofert. Efficiently sampling Archimedean copulas (2009).