General discussion

Archimedean copulas are a large class of copulas that are defined as : ... [ details needed...]

Adding a new ArchimedeanCopula is very easy. The Clayton implementation is as short as:

struct ClaytonCopula{d,T} <: ArchimedeanCopula{d}
    θ::T
    function ClaytonCopula(d,θ)
        if θ < -1/(d-1)
            throw(ArgumentError("Theta must be greater than -1/(d-1)"))
        elseif θ == -1/(d-1)
            return WCopula(d)
        elseif θ == 0
            return IndependentCopula(d)
        elseif θ == Inf
            return MCopula(d)
        else
            return new{d,typeof(θ)}(θ)
        end
    end
end
ϕ(  C::ClaytonCopula,      t) = max(1+C.θ*t,zero(t))^(-1/C.θ)
ϕ⁻¹(C::ClaytonCopula,      t) = (t^(-C.θ)-1)/C.θ
τ(C::ClaytonCopula) = ifelse(isfinite(C.θ), C.θ/(C.θ+2), 1)
τ⁻¹(::Type{ClaytonCopula},τ) = ifelse(τ == 1,Inf,2τ/(1-τ))
williamson_dist(C::ClaytonCopula{d,T}) where {d,T} = C.θ >= 0 ? WilliamsonFromFrailty(Distributions.Gamma(1/C.θ,1),d) : ClaytonWilliamsonDistribution(C.θ,d)

Not everything is necessary however, depending on the features you need. Indeed, the Archimedean API is modular:

To sample an archimedean, only ϕ is required. Indeed, the williamson_dist has a generic fallback that uses WilliamsonTransforms.jl for any generator. Note however that providing the williamson_dist yourself if you know it will allow sampling to be an order of magnitude faster: see how in the ClaytonCopula case we provided a sampler that is different for positive parameters (the generator is completely monotonous, and thus the frailty distribution is known, so we used WiliamsonFromFrailty), and the negative dependent cases, where only the CDF is known and implemented in ClaytonWilliamsonDistribution.

To evaluate the cdf and (log-)density in any dimension, only ϕ and ϕ⁻¹ are needed.

Currently, to fit the copula τ⁻¹ is needed as we use the inverse tau moment method. But we plan on also implementing inverse rho and MLE (density needed).

Note that the generator ϕ follows the convention ϕ(0)=1, while others (e.g., https://en.wikipedia.org/wiki/Copula(probabilitytheory)#Archimedean_copulas) use ϕ⁻¹ as the generator.

Copulas.ArchimedeanCopula — Type

ArchimedeanCopula{d}

Abstract type.

[ The description of the APi done here could be more detailled]

The archimedean copulas are all `<:ArchimedeanCopula{d} for some dimension d. This type serves as an internal interface to handle all archimdean copulas at once.

Adding a new ArchimedeanCopula is very easy. The Clayton implementation is as short as:

struct ClaytonCopula{d,T} <: Copulas.ArchimedeanCopula{d}
    θ::T
end
ClaytonCopula(d,θ)            = ClaytonCopula{d,typeof(θ)}(θ)     # Constructor
ϕ(C::ClaytonCopula, t)        = (1+sign(C.θ)*t)^(-1/C.θ)          # Generator
ϕ⁻¹(C::ClaytonCopula,t)       = sign(C.θ)*(t^(-C.θ)-1)            # Inverse Generator
τ(C::ClaytonCopula)           = C.θ/(C.θ+2)                       # θ -> τ
τ⁻¹(::Type{ClaytonCopula},τ)  = 2τ/(1-τ)                          # τ -> θ
williamson_dist(C::ClaytonCopula{d,T}) where {d,T} = WilliamsonFromFrailty(Distributions.Gamma(1/C.θ,1),d) # Radial distribution

The Archimedean API is modular:

To sample an archimedean, only williamson_dist and ϕ are needed.

To evaluate the cdf and (log-)density in any dimension, only ϕ and ϕ⁻¹ are needed.

Currently, to fit the copula τ⁻¹ is needed as we use the inverse tau moment method. But we plan on also implementing inverse rho and MLE (density needed).

Note that the generator ϕ follows the convention ϕ(0)=1, while others (e.g., https://en.wikipedia.org/wiki/Copula(probabilitytheory)#Archimedean_copulas) use ϕ⁻¹ as the generator.

We plan on implementing the Williamson transformations so that radial-dist can be automaticlaly deduced from ϕ and vice versa, if you dont know much about your archimedean family

If you only know the generator of your copula, take a look at WilliamsonCopula that allows to generate automatically the associated williamson distribution. If on the other hand you have a univaraite positive random variable with no atom at zero, then the williamson transform can produce an archimdean copula out of it, with the same constructor.

The WilliamsonCopula concept

In this package, there is the possibility to directly implement, sample, and evaluate the pdf and cdf of an archimedean copula by only providing its generator, in an efficient way.

Copulas.WilliamsonCopula — Type

WilliamsonCopula{d,Tϕ,TX}

Fields: - ϕ::Tϕ – a function representing the archimedean generator. - X::TX – a random variable that represents its williamson d-transform

Constructors

WilliamsonCopula(X::Distributions.UnivariateDistribution, d)
WilliamsonCopula(ϕ::Function, d)
WilliamsonCopula(ϕ::Function, X::Distributions.UnivariateDistribution, d)

The WilliamsonCopula is the barebone Archimedean Copula that directly leverages the Williamson transform and inverse transform (in their greatest generalities), that are implemented in WilliamsonTransforms.jl. You can construct it by providing the Williamson-d-tranform as a (non-negative) random variable X::Distributions.UnivariateDistribution, or by providing the $d$-monotone generator ϕ::Function as a function from $\mathbb R_+$ to $[0,1]$, decreasing and d-monotone. The other component will be computed automatically. You also have the option to provide both, which will probably be faster if you have an analytical expression for both.

About WilliamsonCopula(X::Distributions.UnivariateDistribution, d): For a univariate non-negative random variable $X$, with cumulative distribution function $F$ and an integer $d\ge 2$, the Williamson-d-transform of $X$ is the real function supported on $[0,\infty[$ given by:

\[\phi(t) = 𝒲_{d}(X)(t) = \int_{t}^{\infty} \left(1 - \frac{t}{x}\right)^{d-1} dF(x) = \mathbb E\left( (1 - \frac{t}{X})^{d-1}_+\right) \mathbb 1_{t > 0} + \left(1 - F(0)\right)\mathbb 1_{t <0}\]

This function has several properties:

We have that $\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

These properties makes this function what is called an archimedean generator, able to generate archimedean copulas in dimensions up to $d$.

About WilliamsonCopula(ϕ::Function, d): On the other hand, WilliamsonCopula(ϕ::Function, d) Computes the inverse Williamson d-transform of the d-monotone archimedean generator ϕ.

A $d$-monotone archimedean generator is a function $\phi$ on $\mathbb R_+$ that has these three properties:

$\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

For such a function $\phi$, the inverse Williamson-d-transform of $\phi$ is the cumulative distribution function $F$ of a non-negative random variable $X$, defined by :

\[F(x) = 𝒲_{d}^{-1}(\phi)(x) = 1 - \frac{(-x)^{d-1} \phi_+^{(d-1)}(x)}{k!} - \sum_{k=0}^{d-2} \frac{(-x)^k \phi^{(k)}(x)}{k!}\]

We return this cumulative distribution function in the form of the corresponding random variable <:Distributions.ContinuousUnivariateDistribution from Distributions.jl. You may then compute : - The cdf via Distributions.cdf - The pdf via Distributions.pdf and the logpdf via Distributions.logpdf - Samples from the distribution via rand(X,n)

References: Williamson, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581 McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.