Archimedean Copulas

Details about what are archimedean copulas (on the math level)

Generic Archimedean Copulas

Details about the generic construction of archimedean copulas (in the package), and details on exported methods that corresponds to this class

Missing docstring.

Missing docstring for ArchimedeanCopula. Check Documenter's build log for details.

Implement your own

Explain how easy it is to implement your own archimedean copulas and work with them. methods:

struct MyAC{d,T} <: ArchimedeanCopula{d}
    par::T
end
ϕ(C::MyAC{d},x)
ϕ⁻¹(C::MyAC{d},x)
τ(C::MyAC{d})
τ⁻¹(::MyAC{d},τ)
radial_dist(C::MyAC{d})

Available Archimedean copulas

Independence

Copulas.IndependentCopula — Type

IndependentCopula{d,T}

Constructor

IndependentCopula(d, θ)

The Independant Copula in dimension $d$ is the simplest copula, that has the form :

\[C(\mathbf{x}) = \prod_{i=1}^d x_i.\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = \exp{-t}\]

Clayton

Copulas.ClaytonCopula — Type

ClaytonCopula{d,T}

Fields:

θ::Real - parameter

Constructor

ClaytonCopula(d, θ)

The Clayton copula in dimension $d$ is parameterized by $\theta \in [-1,\infty)$ when $d=2$ and $\theta \in [0,\infty)$ if $d>2$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

Frank

Copulas.FrankCopula — Type

FrankCopula{d,T}

Fields:

θ::Real - parameter

Constructor

FrankCopula(d, θ)

The Frank copula in dimension $d$ is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

Gumbel

Copulas.GumbelCopula — Type

GumbelCopula{d,T}

Fields:

θ::Real - parameter

Constructor

GumbelCopula(d, θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

Ali-Mikhail-Haq

Copulas.AMHCopula — Type

AMHCopula{d,T}

Fields:

θ::Real - parameter

Constructor

AMHCopula(d, θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [0,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

Joe

Copulas.JoeCopula — Type

JoeCopula{d,T}

Fields:

θ::Real - parameter

Constructor

JoeCopula(d, θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]