Available models

GaussianCopula

Copulas.GaussianCopula — Type

GaussianCopula{d,MT}

Fields:

Σ::MT - covariance matrix

Constructor

GaussianCopula(Σ)

The Gaussian Copula is the copula of a Multivariate normal distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{\Sigma}) = F_{\Sigma}(F_{\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{\Sigma}$ is a cdf of a gaussina random vector and F_{\Sigma,i} is the ith marignal cdf, while `$\Sigma$ is the covariance matrix.

It can be constructed in Julia via:

C = GaussianCopula(Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(GaussianCopula,data)

TCopula

Copulas.TCopula — Type

TCopula{d,MT}

Fields:

df::Int - number of degree of freedom

Σ::MT - covariance matrix

Constructor

TCopula(df,Σ)

The Student's T Copula is the copula of a Multivariate Student distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{n,\Sigma}) = F_{n,\Sigma}(F_{n,\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{n,\Sigma}$ is a cdf of a multivariate student random vector with covariance matrix $\Sigma$ and $n$ degrees of freedom. and F_{n,\Sigma,i} is the ith marignal cdf.

It can be constructed in Julia via:

C = TCopula(n,Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(TCopula,data)

Except that currently it does not work since fit(Distributions.MvTDist,data) does not dispatch.