Extreme Values Tails

MTail

Copulas.MTail — Type

MTail

Corresponds to the MCopula viewed as an etreme value copula.

NoTail

Copulas.NoTail — Type

NoTail

Corresponds to the case where the pickads function is identically One, which means no particular tail behavior.

AsymGalambosTail

Copulas.AsymGalambosTail — Type

AsymGalambosTail{T}

Fields:

α::Real — dependence parameter

θ₁::Real — asymmetry weight in [0,1]

θ₂::Real — asymmetry weight in [0,1]

Constructor

AsymGalambosCopula(α, θ)         # θ as a 2-vector/tuple
ExtremeValueCopula(2, AsymGalambosTail(α, θ))

The (bivariate) asymmetric Galambos extreme–value copula is parameterized by $\alpha \in [0, \infty) and$\theta1, \theta2 \in [0,1]``. It is an EV copula with Pickands function

\[A(t) = 1 - \Big((\theta_1 t)^{-\alpha} + (\theta_2(1-t))^{-\alpha}\Big)^{-1/\alpha},\quad t\in[0,1].\]

Special cases:

α = 0 ⇒ IndependentCopula

θ₁ = θ₂ = 0 ⇒ IndependentCopula

θ₁ = θ₂ = 1 ⇒ GalambosCopula

References:

[61] Families of min-stable multivariate exponential and multivariate extreme value distributions. Statist. Probab, 1990.

AsymLogTail

Copulas.AsymLogTail — Type

AsymLogTail{T}

Fields: - α::Real — dependence parameter (α ≥ 1) - θ₁::Real — asymmetry weight in [0,1] - θ₂::Real — asymmetry weight in [0,1]

Constructor

AsymLogCopula(α, (θ₁, θ₂))
ExtremeValueCopula(2, AsymLogTail(α, (θ₁, θ₂)))

The (bivariate) asymmetric logistic extreme–value copula is parameterized by α ∈ [1, ∞) and θ₁, θ₂ ∈ [0,1]. Its Pickands dependence function is

\[A(t) = \Big( \theta_1^{\alpha}(1-t)^{\alpha} + \theta_2^{\alpha}t^{\alpha} \Big)^{1/\alpha} + (\theta_1 - \theta_2)\,t + 1 - \theta_1, \quad t\in[0,1].\]

Special cases:

θ₁ = θ₂ = 1 ⇒ symmetric Logistic (Gumbel) copula

θ₁ = θ₂ = 0 ⇒ independence (A(t) ≡ 1)

References:

[62] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

AsymMixedTail

Copulas.AsymMixedTail — Type

AsymMixedTail{T}

Fields:

θ₁::Real — parameter

θ₂::Real — parameter

Constructor

AsymMixedCopula((θ₁, θ₂)) ExtremeValueCopula(2, AsymMixedTail((θ₁, θ₂)))

The (bivariate) asymmetric Mixed extreme-value copula is parameterized by two parameters $\theta_1$, $\theta_2$ subject to the following constraints:

θ₁ ≥ 0

θ₁ + θ₂ ≤ 1

θ₁ + 2θ₂ ≤ 1

θ₁ + 3θ₂ ≥ 0

Its Pickands dependence function is

\[A(t) = \theta_{2}t^3 + \theta_{1}t^2 - (\theta_1+\theta_2)t + 1,\quad t\in[0,1].\]

Special cases:

θ₁ = θ₂ = 0 ⇒ IndependentCopula

θ₂ = 0 ⇒ symmetric Mixed copula

References:

[62] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

BC2Tail

Copulas.BC2Tail — Type

BC2Tail{T}

Fields:

a::Real — parameter (a ∈ [0,1])

b::Real — parameter (b ∈ [0,1])

Constructor

BC2Copula(a, b)
ExtremeValueCopula(2, BC2Tail(a, b))

The bivariate BC2 extreme-value copula is parameterized by two parameters $a, b \in [0,1]$. Its Pickands dependence function is

\[A(t) = \max\{a t, b (1-t) \} + \max\{(1-a)t, (1-b)(1-t)\}, \quad t \in [0,1].\]

References:

[63] Mai, J. F., & Scherer, M. (2011). Bivariate extreme-value copulas with discrete Pickands dependence measure. Extremes, 14, 311-324. Springer, 2011.

CuadrasAugeTail

Copulas.CuadrasAugeTail — Type

CuadrasAugeTail{T}

Fields:

θ::Real — dependence parameter, θ ∈ [0,1]

Constructor

CuadrasAugeCopula(θ)
ExtremeValueCopula(2, CuadrasAugeTail(θ))

The (bivariate) Cuadras-Augé extreme-value copula is parameterized by $\theta \in [0,1]$. Its Pickands dependence function is

\[A(t) = \max\{t, 1-t\} + (1-\theta)\min\{t,1-t\}, \quad t \in [0,1].\]

Special cases:

θ = 0 ⇒ IndependentCopula

θ = 1 ⇒ MCopula (comonotone copula)

References:

[45] Mai, J. F., & Scherer, M. (2012). Simulating copulas: stochastic models, sampling algorithms, and applications (Vol. 4). World Scientific.

GalambosTail

Copulas.GalambosTail — Type

GalambosTail{T}

Fields:

θ::Real — dependence parameter, θ ≥ 0

Constructor

GalambosCopula(θ)
ExtremeValueCopula(2, GalambosTail(θ))

The (bivariate) Galambos extreme-value copula is parameterized by $\theta \in [0, \infty)$. Its Pickands dependence function is

\[A(t) = 1 - \Big( t^{-\theta} + (1-t)^{-\theta} \Big)^{-1/\theta}, \quad t \in (0,1).\]

Special cases:

θ = 0 ⇒ IndependentCopula

θ = ∞ ⇒ MCopula (upper Fréchet-Hoeffding bound)

References:

[64] Galambos, J. (1975). Order statistics of samples from multivariate distributions. Journal of the American Statistical Association, 70(351a), 674-680.

HuslerReissTail

Copulas.HuslerReissTail — Type

HuslerReissTail{T}

Fields:

θ::Real — dependence parameter, θ ≥ 0

Constructor

HuslerReissCopula(θ)
ExtremeValueCopula(2, HuslerReissTail(θ))

The (bivariate) Hüsler-Reiss extreme-value copula is parameterized by $\theta \in [0, \infty)$. Its Pickands dependence function is

\[A(t) = t\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{t}{1-t})) +(1-t)\Phi(\theta^{-1}+\frac{1}{2}\theta\log(\frac{1-t}{t}))\]

where $\Phi$ is the standard normal cdf.

Special cases:

θ = 0 ⇒ IndependentCopula

θ = ∞ ⇒ MCopula (upper Fréchet-Hoeffding bound)

References:

[65] Hüsler, J., & Reiss, R. D. (1989). Maxima of normal random vectors: between independence and complete dependence. Statistics & Probability Letters, 7(4), 283-286.

LogTail

Copulas.LogTail — Type

LogTail{T}

Fields:

θ::Real — dependence parameter, θ ∈ [0,1]

Constructor

LogCopula(θ)
ExtremeValueCopula(2, LogTail(θ))

The (bivariate) Mixed extreme-value copula is parameterized by $\theta \in [0,1]$. Its Pickands dependence function is

\[A(t) = \theta t^2 - \theta t + 1, \quad t \in [0,1].\]

Special cases:

θ = 0 ⇒ IndependentCopula

References:

[62] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

MixedTail

Copulas.MixedTail — Type

MixedTail{T}

Fields:

θ::Real — dependence parameter, θ ∈ [0,1]

Constructor

MixedCopula(θ)
ExtremeValueCopula(2, MixedTail(θ))

The (bivariate) Mixed extreme-value copula is parameterized by $\theta \in [0,1]$. Its Pickands dependence function is

\[A(t) = \theta t^2 - \theta t + 1, \quad t \in [0,1].\]

Special cases:

θ = 0 ⇒ IndependentCopula

References:

[62] : Tawn, Jonathan A. "Bivariate extreme value theory: models and estimation." Biometrika 75.3 (1988): 397-415.

MOTail

Copulas.MOTail — Type

MOTail{T}

Fields:

λ₁::Real — parameter ≥ 0

λ₂::Real — parameter ≥ 0

λ₁₂::Real — parameter ≥ 0

Constructor

MOCopula(λ₁, λ₂, λ₁₂)
ExtremeValueCopula(2, MOTail(λ₁, λ₂, λ₁₂))

The (bivariate) Marshall-Olkin extreme-value copula is parameterized by $\lambda_i \in [0,\infty), i = 1, 2, \{1,2\}$ Its Pickands dependence function is

\[A(t) = \frac{\lambda_1 (1-t)}{\lambda_1 + \lambda_{1,2}} + \frac{\lambda_2 t}{\lambda_2 + \lambda_{1,2}} + \lambda_{1,2}\max\left \{\frac{1-t}{\lambda_1 + \lambda_{1,2}}, \frac{t}{\lambda_2 + \lambda_{1,2}} \right \}\]

Special cases:

If λ₁₂ = 0, reduces to an asymmetric independence-like form.

If λ₁ = λ₂ = 0, degenerates to complete dependence.

References:

[45] Mai, J. F., & Scherer, M. (2012). Simulating copulas: stochastic models, sampling algorithms, and applications (Vol. 4). World Scientific.

tEVTail

Copulas.tEVTail — Type

tEVTail{Tdf,Tρ}

Fields:

ν::Real — degrees of freedom (ν > 0)

ρ::Real — correlation parameter (ρ ∈ (-1,1])

Constructor

tEVCopula(ν, ρ)
ExtremeValueCopula(2, tEVTail(ν, ρ))

The (bivariate) extreme-t copula is parameterized by $\nu > 0$ and \rho \in (-1,1]``. Its Pickands dependence function is

\[A(x) = xt_{\nu+1}(Z_x) +(1-x)t_{\nu+1}(Z_{1-x})\]

Where $t_{\nu + 1}$ is the cumulative distribution function (CDF) of the standard t distribution with $\nu + 1$ degrees of freedom and

\[Z_x = \frac{(1+\nu)^{1/2}{\sqrt{1-\theta^2}}\left [ \left (\frac{x}{1-x} \right )^{1/\nu} - \theta \right ]\]

Special cases:

ρ = 0 ⇒ IndependentCopula

ρ = 1 ⇒ M Copula (upper Fréchet-Hoeffding bound)

References:

[66] Nikoloulopoulos, A. K., Joe, H., & Li, H. (2009). Extreme value properties of multivariate t copulas. Extremes, 12, 129-148.

[45]

J.-F. Mai and M. Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 4 (World Scientific, 2012).

[61]

H. Joe. Families of min-stable multivariate exponential and multivariate extreme value distributions. Statistics & probability letters 9, 75–81 (1990).

[62]

J. A. Tawn. Bivariate extreme value theory: models and estimation. Biometrika 75, 397–415 (1988).

[63]

J.-F. Mai and M. Scherer. Bivariate extreme-value copulas with discrete Pickands dependence measure. Extremes 14, 311–324 (2011).

[64]

J. Galambos. Order statistics of samples from multivariate distributions. Journal of the American Statistical Association 70, 674–680 (1975).

[65]

J. Hüsler and R.-D. Reiss. Maxima of normal random vectors: between independence and complete dependence. Statistics & Probability Letters 7, 283–286 (1989).

[66]

A. K. Nikoloulopoulos, H. Joe and H. Li. Extreme value properties of multivariate t copulas. Extremes 12, 129–148 (2009).