Reference

All docstrings are copied here as a general reference.

Copulas.AMHGenerator — Type

AMHGenerator{T}

Fields:

θ::Real - parameter

Constructor

AMHGenerator(θ)
AMHCopula(d,θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [-1,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

Copulas.ArchimedeanCopula — Type

ArchimedeanCopula{d, TG}

Fields: - G::TG : the generator <: Generator.

Constructor:

ArchimedeanCopula(d::Int,G::Generator)

For some archimedean generator G::Generator and some dimenson d, this class models the archimedean copula wich has this generator. More formally, a $d$-monotone archimedean generator is a function $\phi$ on $\mathbb R_+$ that has these three properties:

$\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

In the code, this function is implemented as ϕ(G::Generator,t), and its max monotony d is given by max_monotony(G). The corresponding d-variate archimedean copula writes:

\[C(\mathbf u) = \phi^{-1}\left(\sum_{i=1}^d \phi(u_i)\right)\]

It can be sampled through A Radial simplex decomposition as follows: If $\mathbf U \sim C$, then $U \equal R \mathbf S$ for a random vector $\mathbf S \sim$ Dirichlet(ones(d)), that is uniformity on the d-variate simplex, and a non-negative random variable $R$ that is the Williamson d-transform of \phi. The density, cdf, kendall tau, and may other properties of the Archimedean copula are then directly deduced from this relationship.

There exists several known parametric generators that are implement in the package. For every NamedGenerator <: Generator implemented in the package, we provide a type alias `NamedCopula{d,...} = ArchimedeanCopula{d,NamedGenerator{...}} to be able to manipulate the archimedean copulas without too much hassle for known and usefull special cases.

A generic archimdeanc copula can be constructed as follows:

struct MyGenerator <: Generator end
ϕ(G::MyGenerator,t) = exp(-t) # your archimedean generator, can be any d-monotonous function.
max_monotony(G::MyGenerator) = Inf # could depend on generators parameters. 
C = ArchimedeanCopula(d,MyGenerator())

The obtained model can be used as follows:

spl = rand(C,1000)   # sampling
cdf(C,spl)           # cdf
pdf(C,spl)           # pdf
loglikelihood(C,spl) # llh

The following functions have defaults but can be overridden for performance:

ϕ⁻¹(C::MyArchimedean, t)        # Inverse of ϕ
ϕ⁽¹⁾(C::MyArchimedean, t)       # first defrivative of ϕ
ϕ⁽ᵈ⁾(C::MyArchimedean,t)        # dth defrivative of ϕ
τ(C::MyArchimedean)             # Kendall tau
τ⁻¹(::Type{MyArchimedean},τ) =  # Inverse kendall tau
fit(::Type{MyArchimedean},data) # fitting.

As a bonus, If you know the williamson d-transform of your generator and not your generator itself, you may take a look at WilliamsonGenerator that implements them. If you rather know the frailty distribution, take a look at WilliamsonFromFrailty.

References: Williamson, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581 McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.

Copulas.ClaytonGenerator — Type

ClaytonGenerator{T}

Fields:

θ::Real - parameter

Constructor

ClaytonGenerator(θ)
ClaytonCopula(d,θ)

The Clayton copula in dimension $d$ is parameterized by $\theta \in [-1/(d-1),\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

It has a few special cases:

When θ = -1/(d-1), it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.EllipticalCopula — Type

EllipticalCopula{d,MT}

Abstract type.

[The description of the API should be here...]

Copulas.EmpiricalCopula — Type

EmpiricalCopula{d,MT}

Fields:

u::MT - the matrix of observations.

Constructor

EmpiricalCopula(u;pseudos=true)

The EmpiricalCopula in dimension $d$ is parameterized by a pseudo-data matrix wich should have shape (d,N). Its expression is given as :

\[C(\mathbf x) = \frac{1}{N}\sum_{i=1}^n \mathbf 1_{\mathbf u_i \le \mathbf x}\]

This function is very practical, be be aware that this is not a true copula (since $\mathbf u$ are only pseudo-observations). The constructor allows you to pass dirctly pseudo-observations (the default) or will compute them for you. You can then compute the cdf of the copula, and sample it through the standard interface.

Copulas.FGMCopula — Type

FGMCopula{d,T}

Fields:

θ::Real - parameter

Constructor

FGMCopula(d, θ)

The Multivariate Farlie-Gumbel-Morgenstern (FGM) copula of dimension d has $2^d-d-1$ parameters $\theta$ and function

\[C(\boldsymbol{u})=\prod_{i=1}^{d}u_i \left[1+ \sum_{k=2}^{d}\sum_{1 \leq j_1 < \cdots < j_k \leq d} \theta_{j_1 \cdots j_k} \bar{u}_{j_1}\cdots \bar{u}_{j_k} \right],\]

where $\bar{u}=1-u$.

More details about Farlie-Gumbel-Morgenstern (FGM) copula are found in :

Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.38.

We use the stochastic representation of the copula to obtain random samples.

Blier-Wong, C., Cossette, H., & Marceau, E. (2022). Stochastic representation of FGM copulas using multivariate Bernoulli random variables. Computational Statistics & Data Analysis, 173, 107506.

It has a few special cases:

When d=2 and θ = 0, it is the IndependentCopula.

Copulas.FrankGenerator — Type

FrankGenerator{T}

Fields:

θ::Real - parameter

Constructor

FrankGenerator(θ)
FrankCopula(d,θ)

The Frank copula in dimension $d$ is parameterized by $\theta \in [-\infty,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

It has a few special cases:

When θ = -∞, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.GaussianCopula — Type

GaussianCopula{d,MT}

Fields:

Σ::MT - covariance matrix

Constructor

GaussianCopula(Σ)

The Gaussian Copula is the copula of a Multivariate normal distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{\Sigma}) = F_{\Sigma}(F_{\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{\Sigma}$ is a cdf of a gaussina random vector and F_{\Sigma,i} is the ith marignal cdf, while `$\Sigma$ is the covariance matrix.

It can be constructed in Julia via:

C = GaussianCopula(Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(GaussianCopula,data)

Copulas.GumbelBarnettGenerator — Type

GumbelBarnettGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelBarnettGenerator(θ)
GumbelBarnettCopula(d,θ)

The Gumbel-Barnett copula is an archimdean copula with generator:

\[\phi(t) = \exp{θ^{-1}(1-e^{t})}, 0 \leq \theta \leq 1.\]

More details about Gumbel-Barnett copula are found in:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.437

It has a few special cases:

When θ = 0, it is the IndependentCopula

Copulas.GumbelGenerator — Type

GumbelGenerator{T}

Fields:

θ::Real - parameter

Constructor

GumbelGenerator(θ)
GumbelCopula(d,θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.IndependentGenerator — Type

IndependentGenerator

Constructor

IndependentGenerator()
IndependentCopula(d)

The Independent Copula in dimension $d$ is the simplest copula, that has the form :

\[C(\mathbf{x}) = \prod_{i=1}^d x_i.\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = \exp{-t}\]

Copulas.InvGaussianGenerator — Type

InvGaussianGenerator{T}

Fields:

θ::Real - parameter

Constructor

InvGaussianGenerator(θ)
InvGaussianCopula(d,θ)

The Inverse Gaussian copula in dimension $d$ is parameterized by $\theta \in [0,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{\frac{(1-\sqrt{1+2θ^{2}t}}{θ}}.\]

More details about Inverse Gaussian Archimedean copula are found in :

Mai, Jan-Frederik, and Matthias Scherer. Simulating copulas: stochastic models, sampling algorithms, and applications. Vol. 6. # N/A, 2017. Page 74.

It has a few special cases:

When θ = 0, it is the IndependentCopula

Copulas.JoeGenerator — Type

JoeGenerator{T}

Fields:

θ::Real - parameter

Constructor

JoeGenerator(θ)
JoeCopula(d,θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.MGenerator — Type

MGenerator

Constructor

MGenerator()
MCopula(d)

The Upper Frechet-Hoeffding bound is the copula with the greatest value among all copulas. It correspond to comonotone random vectors.

For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas.

Copulas.PlackettCopula — Type

PlackettCopula{P}

Fields: - θ::Real - parameter

Constructor

PlackettCopula(θ)

Parameterized by $\theta > 0$ The Plackett copula is

\[C_{\theta}(u,v) = \frac{\left [1+(\theta-1)(u+v)\right]- \sqrt{[1+(\theta-1)(u+v)]^2-4uv\theta(\theta-1)}}{2(\theta-1)}\]

and for $\theta = 1$

\[C_{1}(u,v) = uv \]

It has a few special cases:

When θ = 0, is is the MCopula (Upper Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the WCopula (Lower Frechet-Hoeffding bound)

More details about Plackett Copula are found in:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.164

Details about the algorithm to generate copula samples can be seen in the following references:

Johnson, Mark E. Multivariate statistical simulation:
A guide to selecting and generating continuous multivariate distributions.
Vol. 192. John Wiley & Sons, 1987. Page 193.
Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.38.

Copulas.SklarDist — Type

SklarDist{CT,TplMargins}

Fields:

C::CT - The copula

m::TplMargins - a Tuple representing the marginal distributions

Constructor

SklarDist(C,m)

This function allows to construct a random vector specified, through the Sklar Theorem, by its marginals and its copula separately. See Sklar's theorem.

The obtain random vector follows Distributions.jl's API and can be sampled, pdf and cdf can be evaluated, etc... We even provide a fit function. See the folowing exemple code :

using Copulas, Distributions, Random
X₁ = Gamma(2,3)
X₂ = Pareto()
X₃ = LogNormal(0,1)
C = ClaytonCopula(3,0.7) # A 3-variate Clayton Copula with θ = 0.7
D = SklarDist(C,(X₁,X₂,X₃)) # The final distribution

# This generates a (3,1000)-sized dataset from the multivariate distribution D
simu = rand(D,1000)

# While the following estimates the parameters of the model from a dataset: 
D̂ = fit(SklarDist{FrankCopula,Tuple{Gamma,Normal,LogNormal}}, simu)
# Increase the number of observations to get a beter fit (or not?)  

Copulas.SurvivalCopula — Type

SurvivalCopula{d,CT,VI}

Constructor

SurvivalCopula(C,indices)

Computes the survival version of any copula on given indices. From a copula $C$ in dimension $d$, and some indices $i_1,...i_k$ in ${1,...,d}$, the survival copula associated simply reverses its arguments on chosen indices. For exemple, for $d=4$ and indices $(2,3)$, we have:

\[S(u_1,...u_4) = C(u_1,1-u_2,1-u3,u_4)\]

This constructor allows to derive new "survival" families. For exemple, in bivariate cases, this allows to do "rotations". The obtained models can be treated as the starting one, i.e. as a random vector in [0,1]^d with uniforms marginals.

Copulas.TCopula — Type

TCopula{d,MT}

Fields:

df::Int - number of degree of freedom

Σ::MT - covariance matrix

Constructor

TCopula(df,Σ)

The Student's T Copula is the copula of a Multivariate Student distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{n,\Sigma}) = F_{n,\Sigma}(F_{n,\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{n,\Sigma}$ is a cdf of a multivariate student random vector with covariance matrix $\Sigma$ and $n$ degrees of freedom. and F_{n,\Sigma,i} is the ith marignal cdf.

It can be constructed in Julia via:

C = TCopula(n,Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(TCopula,data)

Except that currently it does not work since fit(Distributions.MvTDist,data) does not dispatch.

Copulas.WGenerator — Type

WGenerator

Constructor

WGenerator()
WCopula(d)

The Lower Frechet-Hoeffding bound is the copula with the lowest value among all copulas. Note that $W$ is only a proper copula when $d=2$, in greater dimensions it is still the (pointwise) lower bound, but not a copula anymore. For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas.

Copulas.WilliamsonGenerator — Type

WilliamsonGenerator{TX}

Fields: - X::TX – a random variable that represents its williamson d-transform - d::Int – the dimension of the transformation.

Constructors

WilliamsonGenerator(X::Distributions.UnivariateDistribution, d)

The WilliamsonGenerator allows to construct a d-monotonous archimedean generator from a positive random variable X::Distributions.UnivariateDistribution. The transformation, wich is called the inverse williamson transformation, is implemented in WilliamsonTransforms.jl.

For a univariate non-negative random variable $X$, with cumulative distribution function $F$ and an integer $d\ge 2$, the Williamson-d-transform of $X$ is the real function supported on $[0,\infty[$ given by:

\[\phi(t) = 𝒲_{d}(X)(t) = \int_{t}^{\infty} \left(1 - \frac{t}{x}\right)^{d-1} dF(x) = \mathbb E\left( (1 - \frac{t}{X})^{d-1}_+\right) \mathbb 1_{t > 0} + \left(1 - F(0)\right)\mathbb 1_{t <0}\]

This function has several properties:

We have that $\phi(0) = 1$ and $\phi(Inf) = 0$

$\phi$ is $d-2$ times derivable, and the signs of its derivatives alternates : $\forall k \in 0,...,d-2, (-1)^k \phi^{(k)} \ge 0$.

$\phi^{(d-2)}$ is convex.

These properties makes this function what is called a d-monotone archimedean generator, able to generate archimedean copulas in dimensions up to $d$. Our implementation provides this through the Generator interface: the function $\phi$ can be accessed by

G = WilliamsonGenerator(X, d)
ϕ(G,t)

Note that you'll always have:

max_monotony(WilliamsonGenerator(X,d)) === d

References: Williamson, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581 McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.