Other Copulas

A few copulas, while necessary in certain cases and really useful, are hard to classify. We gather them here for simplicity.

PlackettCopula

Copulas.PlackettCopula — Type

PlackettCopula{P}

Fields: - θ::Real - parameter

Constructor

PlackettCopula(θ)

Parameterized by $\theta > 0$ The Plackett copula is

\[C_{\theta}(u,v) = \frac{\left [1+(\theta-1)(u+v)\right]- \sqrt{[1+(\theta-1)(u+v)]^2-4uv\theta(\theta-1)}}{2(\theta-1)}\]

and for $\theta = 1$

\[C_{1}(u,v) = uv \]

It has a few special cases:

When θ = 0, is is the MCopula (Upper Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the WCopula (Lower Frechet-Hoeffding bound)

More details about Plackett Copula are found in:

Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.164

Details about the algorithm to generate copula samples can be seen in the following references:

Johnson, Mark E. Multivariate statistical simulation:
A guide to selecting and generating continuous multivariate distributions.
Vol. 192. John Wiley & Sons, 1987. Page 193.
Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.38.

FGMCopula

Farlie-Gumbel-Morgenstern (FGM) copula

Copulas.FGMCopula — Type

FGMCopula{d,T}

Fields:

θ::Real - parameter

Constructor

FGMCopula(d, θ)

The Multivariate Farlie-Gumbel-Morgenstern (FGM) copula of dimension d has $2^d-d-1$ parameters $\theta$ and function

\[C(\boldsymbol{u})=\prod_{i=1}^{d}u_i \left[1+ \sum_{k=2}^{d}\sum_{1 \leq j_1 < \cdots < j_k \leq d} \theta_{j_1 \cdots j_k} \bar{u}_{j_1}\cdots \bar{u}_{j_k} \right],\]

where $\bar{u}=1-u$.

More details about Farlie-Gumbel-Morgenstern (FGM) copula are found in :

Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.38.

We use the stochastic representation of the copula to obtain random samples.

Blier-Wong, C., Cossette, H., & Marceau, E. (2022). Stochastic representation of FGM copulas using multivariate Bernoulli random variables. Computational Statistics & Data Analysis, 173, 107506.

It has a few special cases:

When d=2 and θ = 0, it is the IndependentCopula.

Raftery copula

Copulas.RafteryCopula — Type

RafteryCopula{d, P}

Fields: - θ::Real - parameter

Constructor

RafteryCopula(d, θ)

The Multivariate Raftery Copula of dimension d is arameterized by $\theta \in [0,1]$

\[C_{\theta}(\mathbf{u}) = u_{(1)} + \frac{(1 - \theta)(1 - d)}{1 - \theta - d} \left(\prod_{j=1}^{d} u_j\right)^{\frac{1}{1-\theta}} - \sum_{i=2}^{d} \frac{\theta(1-\theta)}{(1-\theta-i)(2-\theta-i)} \left(\prod_{j=1}^{i-1}u_{(j)}\right)^{\frac{1}{1-\theta}}u_{(i)}^{\frac{2-\theta-i}{1-\theta}}\]

where $u_{(1)}, \ldots , u_{(d)}$ denote the order statistics of $u_1, \ldots ,u_d$. More details about Multivariate Raftery Copula are found in the references below.

It has a few special cases:

When θ = 0, it is the IndependentCopula.

When θ = 1, it is the the Fréchet upper bound

References:

[49] Saali, T., M. Mesfioui, and A. Shabri, 2023: Multivariate Extension of Raftery Copula. Mathematics, 11, 414, https://doi.org/10.3390/math11020414.

[3] Nelsen, Roger B. An introduction to copulas. Springer, 2006. Exercise 3.6.