Reference

All docstrings are copied here as a general reference.

Copulas.Copulas — Module

Copulas

<!– –>

What is this package ?

Warning: This is fairly untested and experimental work and the API might change without notice.

This package brings most standard copula features into native Julia: random number generation, pdf and cdf, fitting, copula-based multivariate distributions through Sklar's theorem, etc., while fully complying with the Distributions.jl API (after all, copulas are distributions functions) in order to provide interoperability with other packages based on this API such as, e.g., Turing.jl.

Usually, people that use and work with copulas turn to R, because of the amazing R package copula. While it is still well maintained and regularly updated, the R package copula is a mixture of obscure, heavily optimized C code and more standard R code, which makes it a complicated code base for readability, extensibility, reliability and maintenance.

This is an attempt to provide a very light, fast, reliable and maintainable copula implementation in native Julia (which means, in particular, floating point type agnostic, i.e. compatibility with BigFloat, DoubleFloats, MultiFloats and other kind of numbers). The two most important exported types are:

Copula: an abstract mother type for all the copulas in the package.

SklarDist: allows construction of a multivariate distribution by specifying the copula and the marginals through Sklar's theorem.

What is already implemented

The API contains random number generation, cdf and pdf evaluation, and the fit function from Distributions.jl. A typical use case might look like this:

using Copulas, Distributions, Random
X₁ = Gamma(2,3)
X₂ = Pareto()
X₃ = LogNormal(0,1)
C = ClaytonCopula(3,0.7) # A 3-variate Clayton Copula with θ = 0.7
D = SklarDist(C,(X₁,X₂,X₃)) # The final distribution

# This generates a (3,1000)-sized dataset from the multivariate distribution D
simu = rand(D,1000)

# While the following estimates the parameters of the model from a dataset: 
D̂ = fit(SklarDist{FrankCopula,Tuple{Gamma,Normal,LogNormal}}, simu)
# Increase the number of observations to get a beter fit (or not?)  

Available copula families are:

GaussianCopula,

TCopula,

ArchimedeanCopula (for any generator),

ClaytonCopula,FrankCopula, AMHCopula, JoeCopula, GumbelCopula as example of the ArchimedeanCopula abstract type, see below,

WCopula and MCopula, which are Fréchet-Hoeffding bounds,

EmpiricalCopula to follow closely a given dataset.

The next ones to be implemented will probably be:

Nested archimedeans (general, with the possibility to nest any family with any family, assuming it is possible, with parameter checks.)

Bernstein copula and more general Beta copula as smoothing of the Empirical copula.

CheckerboardCopula (and more generally PatchworkCopula)

Adding a new ArchimedeanCopula is very easy. The Clayton implementation is as short as:

struct ClaytonCopula{d,T} <: Copulas.ArchimedeanCopula{d}
    θ::T
end
ClaytonCopula(d,θ)            = ClaytonCopula{d,typeof(θ)}(θ)     # Constructor
ϕ(C::ClaytonCopula, t)        = (1+sign(C.θ)*t)^(-1/C.θ)          # Generator
ϕ⁻¹(C::ClaytonCopula,t)       = sign(C.θ)*(t^(-C.θ)-1)            # Inverse Generator
τ(C::ClaytonCopula)           = C.θ/(C.θ+2)                       # θ -> τ
τ⁻¹(::Type{ClaytonCopula},τ)  = 2τ/(1-τ)                          # τ -> θ
radial_dist(C::ClaytonCopula) = Distributions.Gamma(1/C.θ,1)      # Radial distribution

The Archimedean API is modular:

To sample an archimedean, only radial_dist and ϕ are needed.

To evaluate the cdf and (log-)density in any dimension, only ϕ and ϕ⁻¹ are needed.

Currently, to fit the copula τ⁻¹ is needed as we use the inverse tau moment method. But we plan on also implementing inverse rho and MLE (density needed).

Note that the generator ϕ follows the convention ϕ(0)=1, while others (e.g., https://en.wikipedia.org/wiki/Copula(probabilitytheory)#Archimedean_copulas) use ϕ⁻¹ as the generator.

We plan on implementing the Williamson transformations so that radial-dist can be automaticlaly deduced from ϕ and vice versa, if you dont know much about your archimedean family

Dev Roadmap

Urgent things

[ ] Add tests and documentation

Next

[ ] Extensive documentation and tests for the current implementation.

[x] Implement archimedean density generally.

[ ] Docs: show how to implement another archimedean.

[ ] Give the user the choice of fitting method via fit(dist,data; method="MLE") or fit(dist,data; method="itau") or fit(dist,data; method="irho").

[ ] Fitting a generic archimedean : should provide an empirical generator

[ ] Make Archimedean more generic : inputing only radial_dist or only phi shoudl be enough to get pdf, cdf, rand, tau, rho, itau, irho, fit, radial_dist, etc... Williamson d-transform and inverse d-transform should be implemented. The checking of nesting possibility should be done automatically with some rules (is phi_inv \circ phi complementely monotonous ? with obviously shortcut for inter-family nestings.)

Maybe later

[ ] Vines?

[ ] NestedArchimedean and very easy implementation of new archimeean copulas via the radial dist or the phi/invphi + Williamson transform.

[ ] BernsteinCopula and BetaCopula could also be implemented.

[ ] PatchworkCopula and CheckerboardCopula: could be nice things to have :)

[ ] Goodness of fits tests ?

Contributions are welcome

Do not hesitate to open an issue to discuss :)

Copulas.AMHCopula — Type

AMHCopula{d,T}

Fields:

θ::Real - parameter

Constructor

AMHCopula(d, θ)

The AMH copula in dimension $d$ is parameterized by $\theta \in [-1,1)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \frac{1-\theta}{e^{-t}-\theta}\]

It has a few special cases:

When θ = 0, it is the IndependentCopula

Copulas.ArchimedeanCopula — Type

ArchimedeanCopula{d}

Abstract type.

[ The description of the APi done here could be more detailled]

The archimedean copulas are all `<:ArchimedeanCopula{d} for some dimension d. This type serves as an internal interface to handle all archimdean copulas at once.

Adding a new ArchimedeanCopula is very easy. The Clayton implementation is as short as:

struct ClaytonCopula{d,T} <: Copulas.ArchimedeanCopula{d}
    θ::T
end
ClaytonCopula(d,θ)            = ClaytonCopula{d,typeof(θ)}(θ)     # Constructor
ϕ(C::ClaytonCopula, t)        = (1+sign(C.θ)*t)^(-1/C.θ)          # Generator
ϕ⁻¹(C::ClaytonCopula,t)       = sign(C.θ)*(t^(-C.θ)-1)            # Inverse Generator
τ(C::ClaytonCopula)           = C.θ/(C.θ+2)                       # θ -> τ
τ⁻¹(::Type{ClaytonCopula},τ)  = 2τ/(1-τ)                          # τ -> θ
radial_dist(C::ClaytonCopula) = Distributions.Gamma(1/C.θ,1)      # Radial distribution

The Archimedean API is modular:

To sample an archimedean, only radial_dist and ϕ are needed.

To evaluate the cdf and (log-)density in any dimension, only ϕ and ϕ⁻¹ are needed.

Currently, to fit the copula τ⁻¹ is needed as we use the inverse tau moment method. But we plan on also implementing inverse rho and MLE (density needed).

Note that the generator ϕ follows the convention ϕ(0)=1, while others (e.g., https://en.wikipedia.org/wiki/Copula(probabilitytheory)#Archimedean_copulas) use ϕ⁻¹ as the generator.

We plan on implementing the Williamson transformations so that radial-dist can be automaticlaly deduced from ϕ and vice versa, if you dont know much about your archimedean family

Copulas.ClaytonCopula — Type

ClaytonCopula{d,T}

Fields:

θ::Real - parameter

Constructor

ClaytonCopula(d, θ)

The Clayton copula in dimension $d$ is parameterized by $\theta \in [-1/(d-1),\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \left(1+\mathrm{sign}(\theta)*t\right)^{-1\frac{1}{\theta}}\]

It has a few special cases:

When θ = -1/(d-1), it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 0, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.EllipticalCopula — Type

EllipticalCopula{d,MT}

Abstract type.

[The description of the API should be here...]

Copulas.EmpiricalCopula — Type

EmpiricalCopula{d,MT}

Fields:

u::MT - the matrix of observations.

Constructor

EmpiricalCopula(u;pseudos=true)

The EmpiricalCopula in dimension $d$ is parameterized by a pseudo-data matrix wich should have shape (d,N). Its expression is given as :

\[C(\mathbf x) = \frac{1}{N}\sum_{i=1}^n \mathbf 1_{\mathbf u_i \le \mathbf x}\]

This function is very practical, be be aware that this is not a true copula (since $\mathbf u$ are only pseudo-observations). The constructor allows you to pass dirctly pseudo-observations (the default) or will compute them for you. You can then compute the cdf of the copula, and sample it through the standard interface.

Copulas.FrankCopula — Type

FrankCopula{d,T}

Fields:

θ::Real - parameter

Constructor

FrankCopula(d, θ)

The Frank copula in dimension $d$ is parameterized by $\theta \in [-\infty,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = -\frac{\log\left(1+e^{-t}(e^{-\theta-1})\right)}{ heta}\]

It has a few special cases:

When θ = -∞, it is the WCopula (Lower Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.GaussianCopula — Type

GaussianCopula{d,MT}

Fields:

Σ::MT - covariance matrix

Constructor

GaussianCopula(Σ)

The Gaussian Copula is the copula of a Multivariate normal distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{\Sigma}) = F_{\Sigma}(F_{\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{\Sigma}$ is a cdf of a gaussina random vector and F_{\Sigma,i} is the ith marignal cdf, while `$\Sigma$ is the covariance matrix.

It can be constructed in Julia via:

C = GaussianCopula(Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(GaussianCopula,data)

Copulas.GumbelCopula — Type

GumbelCopula{d,T}

Fields:

θ::Real - parameter

Constructor

GumbelCopula(d, θ)

The Gumbel copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = \exp{-t^{\frac{1}{θ}}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.IndependentCopula — Type

IndependentCopula{d,T}

Constructor

IndependentCopula(d, θ)

The Independent Copula in dimension $d$ is the simplest copula, that has the form :

\[C(\mathbf{x}) = \prod_{i=1}^d x_i.\]

It happends to be an Archimedean Copula, with generator :

\[\phi(t) = \exp{-t}\]

Copulas.JoeCopula — Type

JoeCopula{d,T}

Fields:

θ::Real - parameter

Constructor

JoeCopula(d, θ)

The Joe copula in dimension $d$ is parameterized by $\theta \in [1,\infty)$. It is an Archimedean copula with generator :

\[\phi(t) = 1 - \left(1 - e^{-t}\right)^{\frac{1}{\theta}}\]

It has a few special cases:

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the MCopula (Upper Frechet-Hoeffding bound)

Copulas.MCopula — Type

MCopula{d}

Constructor

MCopula(d)

The Upper Frechet-Hoeffding bound is the copula with the greatest value among all copulas. It correspond to comonotone random vectors.

For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas, although this link is not represetned by the hierachy of types in the package for preformances reasons.

Copulas.PlackettCopula — Type

PlackettCopula{P}

Fields: - θ::Real - parameter

Constructor

PlackettCopula(θ)

Parameterized by $\theta > 0$ The Plackett copula is

\[\C_{\theta}(u,v) = \frac{\left [1+(\theta-1)(u+v)\right]- \sqrt{[1+(\theta-1)(u+v)]^2-4uv\theta(\theta-1)}}{2(\theta-1)}\]

and for $\theta = 1$ math \C_{1}(u,v) = uv

More details about Plackett copulation are found in Joe, H. (2014). Dependence modeling with copulas. CRC press, Page.164

It has a few special cases:

When θ = 0, is is the MCopula (Upper Frechet-Hoeffding bound)

When θ = 1, it is the IndependentCopula

When θ = ∞, is is the WCopula (Lower Frechet-Hoeffding bound)

Copulas.SklarDist — Type

SklarDist{CT,TplMargins}

Fields:

C::CT - The copula

m::TplMargins - a Tuple representing the marginal distributions

Constructor

SklarDist(C,m)

This function allows to construct a random vector specified, through the Sklar Theorem, by its marginals and its copula separately. See Sklar's theorem.

The obtain random vector follows Distributions.jl's API and can be sampled, pdf and cdf can be evaluated, etc... We even provide a fit function. See the folowing exemple code :

using Copulas, Distributions, Random
X₁ = Gamma(2,3)
X₂ = Pareto()
X₃ = LogNormal(0,1)
C = ClaytonCopula(3,0.7) # A 3-variate Clayton Copula with θ = 0.7
D = SklarDist(C,(X₁,X₂,X₃)) # The final distribution

# This generates a (3,1000)-sized dataset from the multivariate distribution D
simu = rand(D,1000)

# While the following estimates the parameters of the model from a dataset: 
D̂ = fit(SklarDist{FrankCopula,Tuple{Gamma,Normal,LogNormal}}, simu)
# Increase the number of observations to get a beter fit (or not?)  

Copulas.SurvivalCopula — Type

SurvivalCopula{d,CT,VI}

Constructor

SurvivalCopula(C,indices)

Computes the survival version of any copula on given indices. From a copula $C$ in dimension $d$, and some indices $i_1,...i_k$ in ${1,...,d}$, the survival copula associated simply reverses its arguments on chosen indices. For exemple, for $d=4$ and indices $(2,3)$, we have:

\[S(u_1,...u_4) = C(u_1,1-u_2,1-u3,u_4)\]

This constructor allows to derive new "survival" families. For exemple, in bivariate cases, this allows to do "rotations". The obtained models can be treated as the starting one, i.e. as a random vector in [0,1]^d with uniforms marginals.

Copulas.TCopula — Type

TCopula{d,MT}

Fields:

df::Int - number of degree of freedom

Σ::MT - covariance matrix

Constructor

TCopula(df,Σ)

The Student's T Copula is the copula of a Multivariate Student distribution. It is constructed as :

\[C(\mathbf{x}; \boldsymbol{n,\Sigma}) = F_{n,\Sigma}(F_{n,\Sigma,i}^{-1}(x_i),i\in 1,...d)\]

where $F_{n,\Sigma}$ is a cdf of a multivariate student random vector with covariance matrix $\Sigma$ and $n$ degrees of freedom. and F_{n,\Sigma,i} is the ith marignal cdf.

It can be constructed in Julia via:

C = TCopula(n,Σ)

The random number generation works as expected:

rand(C,1000)
# or
Random.rand!(C,u)

And yo can fit the distribution via :

fit(TCopula,data)

Except that currently it does not work since fit(Distributions.MvTDist,data) does not dispatch.

Copulas.WCopula — Type

WCopula{d}

Constructor

WCopula(d)

The Lower Frechet-Hoeffding bound is the copula with the lowest value among all copulas. Note that $W$ is only a proper copula when $d=1$, in greater dimensions it is still the (pointwise) lower bound, but not a copula anymore. For any copula $C$, if $W$ and $M$ are (respectively) the lower and uppder Frechet-Hoeffding bounds, we have that for all $\mathbf{u} \in [0,1]^d$,

\[W(\mathbf{u}) \le C(\mathbf{u}) \le M(\mathbf{u})\]

The two Frechet-Hoeffding bounds are also Archimedean copulas, although this link is not represetned by the hierachy of types in the package for preformances reasons.